Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240809/합과 평균/sum-and-avg.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    float a, b;
+    double a, b;
     cin >> a >> b;
 
     cout.precision(2);
